// features/collections/schemas.ts
import { routing } from "@/lib/i18n/routing"
import * as z from "zod"

const localesValidation = () => {
    return z.object((Object.fromEntries(routing.locales.map(locale => [locale, z.string().optional()]))))
}

export const collectionSchema = z.object({
    name: localesValidation(),
    attributes: z.array(z.object({
        id: z.number().optional(),
        data: z.any().optional(),
        attribute_id: z.string().optional(),
    })).optional(),
    channels: z.array(z.object({
        channel_id: z.number(),
        is_enabled: z.boolean(),
    })).optional(),
})